fix: 수강 이력 서비스 분리
